wsireg_registration/requirements.txt
# Core registration
wsireg>=0.3.0

# WSI I/O
openslide-python>=3.4.1
tifffile>=2023.1.1

# General
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
